pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devicemanager"
version = "0.1.0"
license = "MIT"
readme = "README.md"
description = ""
authors = [{ name = "Szymon Kozioł", email = "szymonkoz@student.agh.edu.pl" }]
requires-python = ">=3.12"
dependencies = [
    "Django~=5.0.3",
    "djangorestframework~=3.15.1",
    "drf-spectacular~=0.27.2",
    "markdown~=3.6",
    "qrcode~=7.4.2",
    "pillow~=10.3.0",
    "django-colorfield~=0.11.0",
    "django-filter~=24.2",
    "django-environ~=0.11.2",
    "django-lifecycle~=1.2.3",
    "psycopg[binary]~=3.1.18",
    "django-extensions~=3.2.3",
    "bpython~=0.24",
    "pyyaml~=6.0.1",
    "django-jazzmin~=3.0.0",
    "django-phonenumber-field~=7.3.0",
    "phonenumbers~=8.13.35",
    "django-debug-toolbar~=4.3.0",
    "django-allauth~=0.61.1",
    "pymupdf~=1.24.3"
]

[project.optional-dependencies]
dev = [
    "pytest~=8.1.1",
    "pre-commit~=3.7.0",
    "ruff==0.3.7",
    "isort==5.13.2",
    "django-upgrade==1.16.0",
    "factory-boy~=3.3.0",
    "pytest-django~=4.8.0",
]

[tool.isort]
profile = "black"

[tool.pip-tools]
generate_hashes = true

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "devicemanager.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
